Introduce ArrayType traits
