+++ src/Type/JustNullableTypeTrait.php
@@ -3,10 +3,13 @@
 namespace PHPStan\Type;
 
 use PHPStan\TrinaryLogic;
+use PHPStan\Type\Traits\NonArrayTypeTrait;
 use function get_class;
 
 trait JustNullableTypeTrait
 {
+
+	use NonArrayTypeTrait;
 
 	/**
 	 * @return string[]
@@ -50,21 +53,6 @@ trait JustNullableTypeTrait
 	public function traverse(callable $cb): Type
 	{
 		return $this;
-	}
-
-	public function isArray(): TrinaryLogic
-	{
-		return TrinaryLogic::createNo();
-	}
-
-	public function isOversizedArray(): TrinaryLogic
-	{
-		return TrinaryLogic::createNo();
-	}
-
-	public function isList(): TrinaryLogic
-	{
-		return TrinaryLogic::createNo();
 	}
 
 	public function isString(): TrinaryLogic
