+++ src/Type/Traits/MaybeArrayTypeTrait.php
@@ -1,0 +1,25 @@
+<?php declare(strict_types = 1);
+
+namespace PHPStan\Type\Traits;
+
+use PHPStan\TrinaryLogic;
+
+trait MaybeArrayTypeTrait
+{
+
+	public function isArray(): TrinaryLogic
+	{
+		return TrinaryLogic::createMaybe();
+	}
+
+	public function isOversizedArray(): TrinaryLogic
+	{
+		return TrinaryLogic::createMaybe();
+	}
+
+	public function isList(): TrinaryLogic
+	{
+		return TrinaryLogic::createMaybe();
+	}
+
+}
